API for Login, etc

+++ SilverTongue.Services/Dictionary/DictionaryService.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using SilverTongue.Data;
-using SilverTongue.Data.Models;
+﻿using SilverTongue.Data.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,10 +17,11 @@ namespace SilverTongue.Services.Dictionary
         /// </summary>
         /// <param name="note"></param>
         /// <returns></returns>
-        public ServiceResponse<UsersDitctionary> AddNote(UsersDitctionary note)
+        public ServiceResponse<UsersDitctionary> AddNote(UsersDitctionary note, int userId)
         {
             try
             {
+                note.User = _db.Users.Find(userId);
                 _db.UsersDicts.Add(note);
                 _db.SaveChanges();
 
@@ -50,11 +49,21 @@ namespace SilverTongue.Services.Dictionary
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public ServiceResponse<Data.Models.UsersDitctionary> DeleteNote(int id)
+        public ServiceResponse<Data.Models.UsersDitctionary> DeleteNote(string word, int userId)
         {
             try
             {
-                var note = _db.UsersDicts.Find(id);
+                var note = _db.UsersDicts.Where(d => d.Word == word.ToLower()).First();
+                if (note.User.Id != userId)
+                {
+                    return new ServiceResponse<UsersDitctionary>
+                    {
+                        Data = null,
+                        Time = DateTime.UtcNow,
+                        Message = "Not found((",
+                        IsSucces = false
+                    };
+                }
                 _db.UsersDicts.Remove(note);
                 _db.SaveChanges();
                 return new ServiceResponse<UsersDitctionary>
@@ -64,7 +73,7 @@ namespace SilverTongue.Services.Dictionary
                     Message = "Deleted note",
                     IsSucces = true
                 };
-                
+
             }
             catch (Exception e)
             {
@@ -78,7 +87,7 @@ namespace SilverTongue.Services.Dictionary
             }
         }
 
-        
+
         /// <summary>
         /// Возвращает запись по Id
         /// </summary>
@@ -86,7 +95,7 @@ namespace SilverTongue.Services.Dictionary
         /// <returns></returns>
         public List<UsersDitctionary> GetByUserId(int id)
         {
-            return _db.UsersDicts.Where(uds=>uds.User.Id==id).ToList();
+            return _db.UsersDicts.Where(uds => uds.User.Id == id).ToList();
         }
